big refactor reusable command code
